@@ -38,79 +39,80 @@
     "object-assign": "^4.0.1",
     "object-values": "^1.0.0",
     "object.map": "^1.0.0",
-    "prop-types": "^15.5.1",
+    "prop-types": "^15.6.0",
     "raf": "^3.3.2",
-    "react": "^15.6.1",
+    "react": "^16.1.0",
     "react-abstract-autocomplete": "^1.0.0",
-    "react-dnd": "^2.0.2",
-    "react-dnd-html5-backend": "^2.0.0",
-    "react-dom": "^15.6.1",
-    "react-hot-loader": "^3.0.0-beta.6",
-    "react-i18next": "^5.4.1",
+    "react-dnd": "^2.5.4",
+    "react-dnd-html5-backend": "^2.5.4",
+    "react-dom": "^16.1.0",
+    "react-hot-loader": "^3.1.1",
+    "react-i18next": "^6.0.6",
     "react-list": "^0.8.7",
-    "react-list-lazy-load": "^1.0.5",
+    "react-list-lazy-load": "^1.1.0",
     "react-markdown": "^2.4.5",
     "react-redux": "^5.0.6",
     "react-transition-group": "^1.2.0",
-    "recompose": "^0.25.0",
-    "reconnecting-websocket": "^3.2.1",
+    "recompose": "^0.26.0",
+    "reconnecting-websocket": "^3.2.2",
     "redux": "^3.7.2",
     "redux-batched-subscribe": "^0.1.6",
     "redux-logger": "^3.0.0",
     "redux-thunk": "^2.0.1",
     "reselect": "^3.0.0",
-    "router": "^1.1.4",
+    "router": "^1.3.2",
     "screenfull": "^3.3.1",
-    "serve-static": "^1.12.4",
+    "shorten-url": "^1.0.0",
     "splitargs": "0.0.7",
     "trumpet": "^1.7.2",
-    "truncate-url": "^1.0.0",
-    "u-wave-parse-chat-markup": "^2.0.0",
+    "u-wave-parse-chat-markup": "^2.1.0",
     "uwave-tutorial": "soonshiny/uwave-tutorial",
     "whatwg-fetch": "^2.0.0",
     "yamlify": "^0.2.0"
   },
   "devDependencies": {
-    "autoprefixer": "^7.1.2",
+    "autoprefixer": "^7.1.5",
     "babel-core": "^6.26.0",
     "babel-eslint": "^7.1.1",
     "babel-loader": "^7.1.2",
-    "babel-plugin-dynamic-import-node": "^1.0.2",
-    "babel-plugin-dynamic-import-system-import": "^1.0.0",
-    "babel-plugin-istanbul": "^4.1.4",
+    "babel-plugin-dynamic-import-node": "^1.1.0",
+    "babel-plugin-istanbul": "^4.1.5",
     "babel-plugin-transform-decorators-legacy": "^1.3.4",
     "babel-plugin-transform-es2015-modules-commonjs": "^6.26.0",
     "babel-plugin-transform-export-extensions": "^6.22.0",
     "babel-plugin-transform-react-constant-elements": "^6.22.0",
     "babel-plugin-transform-react-inline-elements": "^6.22.0",
-    "babel-plugin-transform-react-remove-prop-types": "^0.4.8",
+    "babel-plugin-transform-react-remove-prop-types": "^0.4.10",
     "babel-plugin-transform-runtime": "^6.22.0",
-    "babel-preset-env": "^1.6.0",
+    "babel-preset-env": "^1.6.1",
     "babel-preset-react": "^6.22.0",
     "babel-preset-stage-2": "^6.22.0",
     "babel-register": "^6.26.0",
-    "chai": "^4.1.1",
-    "chai-enzyme": "^0.8.0",
+    "brotli": "^1.3.2",
+    "chai": "^4.1.2",
+    "chai-enzyme": "^1.0.0-beta.0",
+    "compression-webpack-plugin": "^1.0.1",
     "concat-stream": "^1.6.0",
-    "copy-webpack-plugin": "^4.0.1",
-    "cross-env": "^5.0.5",
-    "css-loader": "^0.28.5",
+    "copy-webpack-plugin": "^4.2.0",
+    "cross-env": "^5.1.0",
+    "css-loader": "^0.28.7",
     "cssnano": "^3.10.0",
     "del": "^3.0.0",
-    "enzyme": "^2.9.1",
-    "eslint": "^4.5.0",
-    "eslint-config-airbnb": "^15.1.0",
-    "eslint-config-airbnb-base": "^11.3.2",
+    "enzyme": "^3.0.0",
+    "enzyme-adapter-react-16": "^1.0.2",
+    "eslint": "^4.9.0",
+    "eslint-config-airbnb": "^16.1.0",
+    "eslint-config-airbnb-base": "^12.1.0",
     "eslint-loader": "^1.9.0",
-    "eslint-plugin-import": "^2.7.0",
-    "eslint-plugin-jsx-a11y": "^5.1.1",
-    "eslint-plugin-react": "^7.3.0",
+    "eslint-plugin-import": "^2.8.0",
+    "eslint-plugin-jsx-a11y": "^6.0.2",
+    "eslint-plugin-react": "^7.4.0",
     "explain-error": "^1.0.3",
-    "express": "^4.15.4",
+    "express": "^4.16.2",
     "extract-loader": "^1.0.1",
     "extract-text-webpack-plugin": "^3.0.0",
-    "fetch-mock": "^5.12.2",
-    "file-loader": "^1.1.4",
+    "fetch-mock": "^5.13.1",
+    "file-loader": "^1.1.5",
     "gulp": "^3.9.1",
     "gulp-babel": "^7.0.0",
     "gulp-newer": "^1.3.0",
@@ -121,32 +123,33 @@
     "html-loader": "^0.5.1",
     "html-webpack-plugin": "^2.30.1",
     "http-proxy-middleware": "^0.17.4",
-    "image-webpack-loader": "^3.2.0",
+    "image-webpack-loader": "^3.4.2",
     "lodash-webpack-plugin": "^0.11.1",
     "loud-rejection": "^1.6.0",
-    "mocha": "^3.5.0",
-    "nodemon": "^1.11.0",
-    "nyc": "^11.1.0",
+    "mocha": "^4.0.1",
+    "nodemon": "^1.12.1",
+    "nyc": "^11.2.1",
     "postcss-cssnext": "^3.0.2",
     "postcss-import": "^11.0.0",
-    "postcss-loader": "^2.0.6",
+    "postcss-loader": "^2.0.8",
     "postcss-url": "^7.1.2",
     "proxyquire": "^1.7.11",
-    "react-test-renderer": "^15.6.1",
+    "react-test-renderer": "^16.0.0",
     "recaptcha-test-keys": "^1.0.0",
-    "run-sequence": "^2.1.0",
+    "run-sequence": "^2.2.0",
     "sinon": "^4.0.0",
     "sinon-chai": "^2.13.0",
-    "style-loader": "^0.18.2",
+    "style-loader": "^0.19.0",
     "through2": "^2.0.3",
     "u-wave-source-soundcloud": "^1.1.0",
     "u-wave-source-youtube": "^1.1.1",
     "u-wave-web-emojione": "^2.0.0",
-    "uglifyjs-webpack-plugin": "^1.0.0-beta.2",
-    "webpack": "^3.5.5",
+    "uglifyjs-webpack-plugin": "^1.0.0-beta.3",
+    "webpack": "^3.8.1",
     "webpack-common-shake": "^1.5.3",
     "webpack-dev-middleware": "^1.12.0",
-    "webpack-hot-middleware": "^2.18.2",
+    "webpack-hot-middleware": "^2.20.0",
+    "webpack-subresource-integrity": "^1.0.1",
     "yaml-loader": "^0.5.0"
   },
   "scripts": {
