@@ -16,7 +16,7 @@ register(
   {
     action: (value) => {
       const volume = parseInt(value, 10);
-      if (!isFinite(volume) || volume < 0 || volume > 100) {
+      if (!Number.isFinite(volume) || volume < 0 || volume > 100) {
         return log('Volume must be a number between 0 and 100.');
       }
       return setVolume(volume);
