@@ -1,5 +1,5 @@
 const React = require('react');
-const renderToStaticMarkup = require('react-dom/server').renderToStaticMarkup;
+const { renderToStaticMarkup } = require('react-dom/server');
 
 const MuiThemeProvider = require('material-ui/styles/MuiThemeProvider').default;
 const getMuiTheme = require('material-ui/styles/getMuiTheme').default;
@@ -12,11 +12,9 @@ module.exports = function renderLoadingScreen() {
 
   muiTheme.userAgent = 'all';
 
-  return renderToStaticMarkup(
-    React.createElement(
-      MuiThemeProvider,
-      { muiTheme: getMuiTheme(muiTheme) },
-      React.createElement(LoadingScreen)
-    )
-  );
+  return renderToStaticMarkup(React.createElement(
+    MuiThemeProvider,
+    { muiTheme: getMuiTheme(muiTheme) },
+    React.createElement(LoadingScreen)
+  ));
 };
